feat(*): create V1 module
