+++ test/jd_pay/service_test.rb
@@ -52,21 +52,7 @@ class JdPay::ServiceTest < Minitest::Test
   end
 
   def test_pc_pay
-    stub_request(:post, JdPay::Service::PC_PAY_URL).with(headers:  {'Content-Type' => 'application/x-www-form-urlencoded'}, body: {
-      version: 'V2.0',
-      merchant: '22294531',
-      tradeTime: 'd9668085c69c2ecb73421627b78590400135be4ad908a2a7',
-      currency: 'ac7132c57f10d3ce',
-      tradeNum: 'b9a9953b73f9f8eb3612d06b873f74e9',
-      tradeName: '25fc4ff3e72f364fc1d80691e88fd7df',
-      amount: 'e5a6c3761ab9ddaf',
-      orderType: 'e00c693e6c5b8a60',
-      notifyUrl: 'd54a8a984359284ec3d58d814e65630300d528555ef1e14ebe9ff87c62d78770',
-      callbackUrl: 'b470f968b7165dcf745e8e129f073434ef9ec43bd64dda9ba2fe77001f4c7347',
-      userId: '29d899fc35dc7aa50640156c06e5ec0f',
-      sign: 'di/QyXcBJGY/Avb8ilTEUsCxto6F5fbjX4mvGt0a0J2XD1dTJ8sPidfTHofagsgz9H/CP3FC6Hqca49woS7o8vkn4oPfA+coHfxBnm4QtBcvdeQjW3fKq2IKbtGG7UGNGXEJc9Gk7BnXNnuvnbEc5FVFcvbqirdNi+7opO9jmaM='
-    }).to_return(headers: {'Location' => 'payCashier?tradeNum=12345678&ourTradeNum=1234567890&key=fookey'}, status: 302)
-
+    html_str = "<html><body onload=document.getElementById('payForm').submit(); style='display: none;'><form action=https://wepay.jd.com/jdpay/saveOrder method='post' id='payForm'><input type='text' name=version value=V2.0><input type='text' name=merchant value=22294531><input type='text' name=tradeTime value=d9668085c69c2ecb73421627b78590400135be4ad908a2a7><input type='text' name=currency value=ac7132c57f10d3ce><input type='text' name=tradeNum value=b9a9953b73f9f8eb3612d06b873f74e9><input type='text' name=tradeName value=25fc4ff3e72f364fc1d80691e88fd7df><input type='text' name=amount value=e5a6c3761ab9ddaf><input type='text' name=orderType value=e00c693e6c5b8a60><input type='text' name=notifyUrl value=d54a8a984359284ec3d58d814e65630300d528555ef1e14ebe9ff87c62d78770><input type='text' name=callbackUrl value=b470f968b7165dcf745e8e129f073434ef9ec43bd64dda9ba2fe77001f4c7347><input type='text' name=userId value=29d899fc35dc7aa50640156c06e5ec0f><input type='text' name=sign value=di/QyXcBJGY/Avb8ilTEUsCxto6F5fbjX4mvGt0a0J2XD1dTJ8sPidfTHofagsgz9H/CP3FC6Hqca49woS7o8vkn4oPfA+coHfxBnm4QtBcvdeQjW3fKq2IKbtGG7UGNGXEJc9Gk7BnXNnuvnbEc5FVFcvbqirdNi+7opO9jmaM=></form></body></html>"
     params = {
       tradeNum: '12345678',
       tradeName: '测试商品',
@@ -77,26 +63,11 @@ class JdPay::ServiceTest < Minitest::Test
       userId: "0000001",
       tradeTime: '20170718101010'
     }
-    assert_equal 'https://wepay.jd.com/jdpay/payCashier?tradeNum=12345678&ourTradeNum=1234567890&key=fookey', JdPay::Service.pc_pay(params)
+    assert_equal html_str, JdPay::Service.pc_pay(params)
   end
 
   def test_h5_pay
-    stub_request(:post, JdPay::Service::H5_PAY_URL).with(headers:  {'Content-Type' => 'application/x-www-form-urlencoded'}, body: {
-      version: 'V2.0',
-      merchant: '22294531',
-      tradeTime: 'd9668085c69c2ecb73421627b78590400135be4ad908a2a7',
-      currency: 'ac7132c57f10d3ce',
-      tradeNum: 'b9a9953b73f9f8eb3612d06b873f74e9',
-      tradeName: '25fc4ff3e72f364fc1d80691e88fd7df',
-      amount: 'e5a6c3761ab9ddaf',
-      orderType: 'e00c693e6c5b8a60',
-      notifyUrl: 'd54a8a984359284ec3d58d814e65630300d528555ef1e14ebe9ff87c62d78770',
-      callbackUrl: 'b470f968b7165dcf745e8e129f073434ef9ec43bd64dda9ba2fe77001f4c7347',
-      userId: '29d899fc35dc7aa50640156c06e5ec0f',
-      sign: 'di/QyXcBJGY/Avb8ilTEUsCxto6F5fbjX4mvGt0a0J2XD1dTJ8sPidfTHofagsgz9H/CP3FC6Hqca49woS7o8vkn4oPfA+coHfxBnm4QtBcvdeQjW3fKq2IKbtGG7UGNGXEJc9Gk7BnXNnuvnbEc5FVFcvbqirdNi+7opO9jmaM='
-    }).to_return(headers: {'Location' => 'payCashier?tradeNum=12345678&ourTradeNum=1234567890&key=fookey'}, status: 302)
-
-
+    html_str = "<html><body onload=document.getElementById('payForm').submit(); style='display: none;'><form action=https://h5pay.jd.com/jdpay/saveOrder method='post' id='payForm'><input type='text' name=version value=V2.0><input type='text' name=merchant value=22294531><input type='text' name=tradeTime value=d9668085c69c2ecb73421627b78590400135be4ad908a2a7><input type='text' name=currency value=ac7132c57f10d3ce><input type='text' name=tradeNum value=b9a9953b73f9f8eb3612d06b873f74e9><input type='text' name=tradeName value=25fc4ff3e72f364fc1d80691e88fd7df><input type='text' name=amount value=e5a6c3761ab9ddaf><input type='text' name=orderType value=e00c693e6c5b8a60><input type='text' name=notifyUrl value=d54a8a984359284ec3d58d814e65630300d528555ef1e14ebe9ff87c62d78770><input type='text' name=callbackUrl value=b470f968b7165dcf745e8e129f073434ef9ec43bd64dda9ba2fe77001f4c7347><input type='text' name=userId value=29d899fc35dc7aa50640156c06e5ec0f><input type='text' name=sign value=di/QyXcBJGY/Avb8ilTEUsCxto6F5fbjX4mvGt0a0J2XD1dTJ8sPidfTHofagsgz9H/CP3FC6Hqca49woS7o8vkn4oPfA+coHfxBnm4QtBcvdeQjW3fKq2IKbtGG7UGNGXEJc9Gk7BnXNnuvnbEc5FVFcvbqirdNi+7opO9jmaM=></form></body></html>"
     params = {
       tradeNum: '12345678',
       tradeName: '测试商品',
@@ -107,7 +78,7 @@ class JdPay::ServiceTest < Minitest::Test
       userId: "0000001",
       tradeTime: '20170718101010'
     }
-    assert_equal 'https://h5pay.jd.com/jdpay/payCashier?tradeNum=12345678&ourTradeNum=1234567890&key=fookey', JdPay::Service.h5_pay(params)
+    assert_equal html_str, JdPay::Service.h5_pay(params)
   end
 
   def test_revoke
@@ -120,8 +91,8 @@ class JdPay::ServiceTest < Minitest::Test
     assert_equal @result, JdPay::Service.revoke(params)
   end
 
-  def test_verify_notification
-    assert_equal @result, JdPay::Service.verify_notification(@response_body)
+  def test_notify_verify
+    assert_equal @result, JdPay::Service.notify_verify(@response_body)
   end
 
   def test_user_relation
@@ -158,58 +129,5 @@ class JdPay::ServiceTest < Minitest::Test
     <?xml version="1.0" encoding="UTF-8"?><jdpay><version>V2.0</version><merchant>22294531</merchant><result><code>000000</code><desc>成功</desc></result><sign>DUUlg3VLAFgx7vT6nCRBagmyJ8O8xsGC70kb6z9FjSO6vy3Vi7VNJ9rYizT+zP4JXOWxyeOAcgpY4O1I5tT1xrmh0N6k/z8PmRbKYXjUNNY999+teh5Ahwy9aigHw0u1ilWFcKmAMhF1gfyjX66WSKxMQASnDHTCEh8m1VBf76o=</sign></jdpay>
     EOF
     assert_equal Hash.from_xml(expect_xml), JdPay::Service.cancel_user({userId: "123"})
-  end
-
-  def test_verify_redirection
-    # invalid param
-    foo_params = {
-      'amount' => '216d686c8900e730',
-      'currency' => 'f42451d04fdb7b06',
-      'note' => '',
-      'sign' => 'hO5dwBI6P9jpcPComYNzXct5P5Y2yHI4awojnp1xh1AaJseqegga5VsTV9GMlypXio8wK4PztvvP
- U4z14YMka0PDVkzuaA9LV7nPwbQQPj+0tT1F7kiVLIhk//Er2XtzNH93CwReKY9YZEP9NIsgaLTb
- dK6JAdQiOQyz6CN7dDY=',
-      'status' => 'f60d28559734b3a0',
-      'tradeNum' => '8deb6f6eb7daed9119c6088b63f8eb3986f651ad81f83ba5',
-      'tradeTime' => 'cdf7da29ab4ff16833fcd7d819ef300774d016b035035df2'
-    }
-
-    begin
-      decrypted_param = JdPay::Service.verify_redirection(foo_params)
-    rescue => e
-      assert_equal e.class, JdPay::Error::InvalidRedirection
-    end
-
-    # valid param
-    params = {
-      'amount' => 'e5a6c3761ab9ddaf',
-      'currency' => 'ac7132c57f10d3ce',
-      'note' => '',
-      'sign' => 'hO5dwBI6P9jpcPComYNzXct5P5Y2yHI4awojnp1xh1AaJseqegga5VsTV9GMlypXio8wK4PztvvP
- U4z14YMka0PDVkzuaA9LV7nPwbQQPj+0tT1F7kiVLIhk//Er2XtzNH93CwReKY9YZEP9NIsgaLTb
- dK6JAdQiOQyz6CN7dDY=',
-      'status' => 'e00c693e6c5b8a60',
-      'tradeNum' => '38442bada8fcc81581d733ff756dc840272bf690d6b0c0e4',
-      'tradeTime' => 'c33390d9b3d6d6699706ea1467616603a2a1506bbaaf0f59'
-    }
-
-    decrypted_param = JdPay::Service.verify_redirection(params)
-    assert_equal decrypted_param, {"amount"=>"1", "currency"=>"CNY", "note"=>"", "status"=>"0", "tradeNum"=>"dev-1531882041-111", "tradeTime"=>"20180718104826"}
-
-
-    # valid param & note is not null
-    new_params = {
-      'amount' => 'e5a6c3761ab9ddaf',
-      'currency' => 'ac7132c57f10d3ce',
-      'note' => '0f0071f565cf49ef82aeee8910e9f2f4',
-      'sign' => 'hqbVC4wf2MLFC7bZZR1mofzc+uLQTNNZiyzpGR/1iq51DcFwrtN4/esOIVVYF2hX/+YNXOL9y5Ii
- i7/6Rs3z/ozROFOzD72t/nvpUI+6gOVjxgAui8bn/K3xkyl0Bo8QgBt16e6KnnRW5QVvJ2jjZRyG
- rzWFH/zgZf06cNaWolY=',
-      'status' => 'e00c693e6c5b8a60',
-      'tradeNum' => '38442bada8fcc815cc6386f24131481a4205f9d165019e8f',
-      'tradeTime' => 'c33390d9b3d6d669fa4bbfc60f925498fe8ac39c72e43a1f'
-    }
-    new_decrypted_param = JdPay::Service.verify_redirection(new_params)
-    assert_equal new_decrypted_param, {"amount"=>"1", "currency"=>"CNY", "note"=>"test product", "status"=>"0", "tradeNum"=>"dev-1531992606-125", "tradeTime"=>"20180719173221"}
   end
 end
